backend/requirements.txt
django==3.2.5
djangorestframework==3.12.4
djoser==2.1.0
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
drf-extra-fields==3.1.1
pillow==8.3.2
reportlab==3.5.68